+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Cert.hs
@@ -77,7 +77,8 @@ import NoThunks.Class (NoThunks)
 
 data CertEnv era = CertEnv
   { cePParams :: !(PParams era)
-  , ceCurrentEpoch :: !EpochNo
+  , ceCurrentEpoch :: EpochNo
+  -- ^ Lazy on purpose, because not all certificates need to know the current EpochNo
   , ceCurrentCommittee :: StrictMaybe (Committee era)
   , ceCommitteeProposals :: Map.Map (GovPurposeId 'CommitteePurpose era) (GovActionState era)
   }