+++ libs/cardano-ledger-core/src/Cardano/Ledger/Slot.hs
@@ -17,6 +17,7 @@ module Cardano.Ledger.Slot (
   EpochInfo,
   -- Block number
   BlockNo (..),
+  epochFromSlot,
   epochInfoEpoch,
   epochInfoFirst,
   epochInfoSize,
@@ -28,9 +29,8 @@ import Cardano.Slotting.Block (BlockNo (..))
 import Cardano.Slotting.EpochInfo (EpochInfo)
 import qualified Cardano.Slotting.EpochInfo as EI
 import Cardano.Slotting.Slot (EpochNo (..), EpochSize (..), SlotNo (..))
-import Control.Monad.Trans (lift)
-import Control.Monad.Trans.Reader (ask)
-import Data.Functor.Identity (Identity)
+import Control.Monad.Reader (Reader, ask, asks)
+import Data.Functor.Identity (Identity, runIdentity)
 import Data.Word (Word64)
 import GHC.Generics (Generic)
 import GHC.Stack (HasCallStack)
@@ -46,7 +46,6 @@ instance Semigroup Duration where
 
 instance Monoid Duration where
   mempty = Duration 0
-  mappend = (<>)
 
 (-*) :: SlotNo -> SlotNo -> Duration
 (SlotNo s) -* (SlotNo t) = Duration (if s > t then s - t else t - s)
@@ -58,26 +57,31 @@ instance Monoid Duration where
 (*-) :: SlotNo -> Duration -> SlotNo
 (SlotNo s) *- (Duration d) = SlotNo (if s > d then s - d else 0)
 
+epochFromSlot :: SlotNo -> Reader Globals EpochNo
+epochFromSlot slot = do
+  ei <- asks epochInfoPure
+  pure $ epochInfoEpoch ei slot
+
 epochInfoEpoch ::
   HasCallStack =>
   EpochInfo Identity ->
   SlotNo ->
-  ShelleyBase EpochNo
-epochInfoEpoch ei = lift . EI.epochInfoEpoch ei
+  EpochNo
+epochInfoEpoch ei = runIdentity . EI.epochInfoEpoch ei
 
 epochInfoFirst ::
   HasCallStack =>
   EpochInfo Identity ->
   EpochNo ->
-  ShelleyBase SlotNo
-epochInfoFirst ei = lift . EI.epochInfoFirst ei
+  SlotNo
+epochInfoFirst ei = runIdentity . EI.epochInfoFirst ei
 
 epochInfoSize ::
   HasCallStack =>
   EpochInfo Identity ->
   EpochNo ->
-  ShelleyBase EpochSize
-epochInfoSize ei = lift . EI.epochInfoSize ei
+  EpochSize
+epochInfoSize ei = runIdentity . EI.epochInfoSize ei
 
 -- | Figure out a slot number that is two stability windows before the end of the next
 -- epoch. Together with the slot number we also return the current epoch number and the
@@ -89,9 +93,9 @@ epochInfoSize ei = lift . EI.epochInfoSize ei
 getTheSlotOfNoReturn :: HasCallStack => SlotNo -> ShelleyBase (EpochNo, SlotNo, EpochNo)
 getTheSlotOfNoReturn slot = do
   globals@Globals {stabilityWindow} <- ask
-  let !epochInfo = epochInfoPure globals
-  epochNo <- epochInfoEpoch epochInfo slot
-  let !nextEpochNo = succ epochNo
-  firstSlotNextEpoch <- epochInfoFirst epochInfo nextEpochNo
-  let !pointOfNoReturn = firstSlotNextEpoch *- Duration (2 * stabilityWindow)
+  let epochInfo = epochInfoPure globals
+      epochNo = epochInfoEpoch epochInfo slot
+      nextEpochNo = succ epochNo
+      firstSlotNextEpoch = epochInfoFirst epochInfo nextEpochNo
+      pointOfNoReturn = firstSlotNextEpoch *- Duration (2 * stabilityWindow)
   pure (epochNo, pointOfNoReturn, nextEpochNo)